Convert C-style casts to C++ style casts.
git-svn-id: http://svn.drobilla.net/lad@2432 a436a847-0d15-0410-975c-d299462d15a1

+++ trunk/flowcanvas/src/Ellipse.cpp
@@ -137,7 +137,7 @@ void
 Ellipse::zoom(double z)
 {
 	if (_label)
-		_label->property_size() = static_cast<int>(floor((double)9000.0f * z));
+		_label->property_size() = static_cast<int>(floor(9000.0 * z));
 }
